// frontend/src/declarations/icp_ledger/icp_ledger.did.d.ts
export interface Account {
	owner: Principal;
	subaccount: [] | [SubAccount];
}
export interface AccountBalanceArgs {
	account: AccountIdentifier;
}
export interface AccountBalanceArgsDfx {
	account: TextAccountIdentifier;
}
export type AccountIdentifier = Uint8Array | number[];
// ...
export type Icrc1Timestamp = bigint;
export type Icrc1Tokens = bigint;
// ...
export type Memo = bigint;
// ...
export type SubAccount = Uint8Array | number[];
export type TextAccountIdentifier = string;
export interface TimeStamp {
	timestamp_nanos: bigint;
}
export interface Tokens {
	e8s: bigint;
}
export interface Transaction {
	memo: Memo;
	icrc1_memo: [] | [Uint8Array | number[]];
	operation: [] | [Operation];
	created_at_time: TimeStamp;
}
export interface TransferArg {
	to: Account;
	fee: [] | [Icrc1Tokens];
	memo: [] | [Uint8Array | number[]];
	from_subaccount: [] | [SubAccount];
	created_at_time: [] | [Icrc1Timestamp];
	amount: Icrc1Tokens;
}
export interface TransferArgs {
	to: AccountIdentifier;
	fee: Tokens;
	memo: Memo;
	from_subaccount: [] | [SubAccount];
	created_at_time: [] | [TimeStamp];
	amount: Tokens;
}